prepwork/test/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.5.2",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
